[Mac] Added support for new key names on menubar
